Move library and author index factories to own modules
