+++ network-api/networkapi/wagtailpages/factory/research_hub/author_index.py
@@ -1,0 +1,13 @@
+import factory
+import wagtail_factories
+
+from networkapi.wagtailpages import models as wagtailpage_models
+from networkapi.wagtailpages.factory import image_factory
+
+
+class ResearchAuthorsIndexPageFactory(wagtail_factories.PageFactory):
+    class Meta:
+        model = wagtailpage_models.ResearchAuthorsIndexPage
+
+    title = "Authors"
+    banner_image = factory.SubFactory(image_factory.ImageFactory)
