+++ network-api/networkapi/wagtailpages/tests/research_hub/base.py
@@ -1,5 +1,6 @@
+from networkapi.wagtailpages.factory.research_hub import author_index as author_index_factory
 from networkapi.wagtailpages.factory.research_hub import landing_page as landing_page_factory
-from networkapi.wagtailpages.factory import research_hub as research_factory
+from networkapi.wagtailpages.factory.research_hub import library_page as library_page_factory
 from networkapi.wagtailpages.tests import base as test_base
 
 
@@ -14,10 +15,10 @@ class ResearchHubTestCase(test_base.WagtailpagesTestCase):
         cls.landing_page = landing_page_factory.ResearchLandingPageFactory(
             parent=homepage,
         )
-        cls.library_page = research_factory.ResearchLibraryPageFactory(
+        cls.library_page = library_page_factory.ResearchLibraryPageFactory(
             parent=cls.landing_page,
         )
-        cls.author_index = research_factory.ResearchAuthorsIndexPageFactory(
+        cls.author_index = author_index_factory.ResearchAuthorsIndexPageFactory(
             parent=cls.landing_page,
             title="Authors",
         )
